work experience completed, change the personal information to one to one inverse

+++ src/main/java/com/portafolio/security/repository/UserRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
-    boolean existsById(int id);
+    boolean existsById(Long id);
 }
